requirements.txt
boto3
psycopg2-binary
pandas
numpy==1.24.4